+++ staticfiles/management/commands/collectstatic.py
@@ -7,8 +7,10 @@ from optparse import make_option
 from django.core.files.storage import FileSystemStorage
 from django.core.management.base import CommandError, NoArgsCommand
 from django.utils.encoding import smart_str, smart_unicode
+from django.utils.datastructures import SortedDict
 
 from staticfiles import finders, storage
+from staticfiles.conf import settings
 
 
 class Command(NoArgsCommand):
@@ -46,12 +48,14 @@ class Command(NoArgsCommand):
             help="Ignore post-processing errors raised on missing file."),
     )
     help = "Collect static files in a single location."
+    requires_model_validation = False
 
     def __init__(self, *args, **kwargs):
         super(NoArgsCommand, self).__init__(*args, **kwargs)
         self.copied_files = []
         self.symlinked_files = []
         self.unmodified_files = []
+        self.post_processed_files = []
         self.storage = storage.staticfiles_storage
         try:
             self.storage.path('')
@@ -63,18 +67,28 @@ class Command(NoArgsCommand):
         if hasattr(os, 'stat_float_times'):
             os.stat_float_times(False)
 
-    def handle_noargs(self, **options):
+    def set_options(self, **options):
+        """
+        Set instance variables based on an options dict
+        """
+        self.interactive = options['interactive']
+        self.verbosity = int(options.get('verbosity', 1))
+        self.symlink = options['link']
         self.clear = options['clear']
         self.dry_run = options['dry_run']
         ignore_patterns = options['ignore_patterns']
+        ignore_patterns.extend(settings.STATICFILES_IGNORE_PATTERNS)
         if options['use_default_ignore_patterns']:
             ignore_patterns += ['CVS', '.*', '*~']
         self.ignore_patterns = list(set(ignore_patterns))
-        self.interactive = options['interactive']
-        self.symlink = options['link']
-        self.verbosity = int(options.get('verbosity', 1))
         self.post_process = options['post_process']
 
+    def collect(self):
+        """
+        Perform the bulk of the work of collectstatic.
+
+        Split off from handle_noargs() to facilitate testing.
+        """
         if self.symlink:
             if sys.platform == 'win32':
                 raise CommandError("Symlinking is not supported by this "
@@ -82,6 +96,46 @@ class Command(NoArgsCommand):
             if not self.local:
                 raise CommandError("Can't symlink to a remote destination.")
 
+        if self.clear:
+            self.clear_dir('')
+
+        if self.symlink:
+            handler = self.link_file
+        else:
+            handler = self.copy_file
+
+        found_files = SortedDict()
+        for finder in finders.get_finders():
+            for path, storage in finder.list(self.ignore_patterns):
+                # Prefix the relative path if the source storage contains it
+                if getattr(storage, 'prefix', None):
+                    prefixed_path = os.path.join(storage.prefix, path)
+                else:
+                    prefixed_path = path
+                found_files[prefixed_path] = (storage, path)
+                handler(path, prefixed_path, storage)
+
+        # Here we check if the storage backend has a post_process
+        # method and pass it the list of modified files.
+        if self.post_process and hasattr(self.storage, 'post_process'):
+            processor = self.storage.post_process(found_files,
+                                                  dry_run=self.dry_run)
+            for original_path, processed_path, processed in processor:
+                if processed:
+                    self.log(u"Post-processed '%s' as '%s" %
+                             (original_path, processed_path), level=1)
+                    self.post_processed_files.append(original_path)
+                else:
+                    self.log(u"Skipped post-processing '%s'" % original_path)
+
+        return {
+            'modified': self.copied_files + self.symlinked_files,
+            'unmodified': self.unmodified_files,
+            'post_processed': self.post_processed_files,
+        }
+
+    def handle_noargs(self, **options):
+        self.set_options(**options)
         # Warn before doing anything more.
         if (isinstance(self.storage, FileSystemStorage) and
                 self.storage.location):
@@ -109,49 +163,25 @@ Type 'yes' to continue, or 'no' to cancel: """
             if confirm != 'yes':
                 raise CommandError("Collecting static files cancelled.")
 
-        if self.clear:
-            self.clear_dir('')
-
-        handler = {
-            True: self.link_file,
-            False: self.copy_file,
-        }[self.symlink]
-
-        found_files = []
-        for finder in finders.get_finders():
-            for path, storage in finder.list(self.ignore_patterns):
-                # Prefix the relative path if the source storage contains it
-                if getattr(storage, 'prefix', None):
-                    prefixed_path = os.path.join(storage.prefix, path)
-                else:
-                    prefixed_path = path
-                found_files.append(prefixed_path)
-                handler(path, prefixed_path, storage)
-
-        # Here we check if the storage backend has a post_process
-        # method and pass it the list of modified files.
-        if self.post_process and hasattr(self.storage, 'post_process'):
-            post_processed = self.storage.post_process(found_files, **options)
-            for path in post_processed:
-                self.log(u"Post-processed '%s'" % path, level=1)
-        else:
-            post_processed = []
-
-        modified_files = self.copied_files + self.symlinked_files
-        actual_count = len(modified_files)
-        unmodified_count = len(self.unmodified_files)
+        collected = self.collect()
+        modified_count = len(collected['modified'])
+        unmodified_count = len(collected['unmodified'])
+        post_processed_count = len(collected['post_processed'])
 
         if self.verbosity >= 1:
-            template = ("\n%(actual_count)s %(identifier)s %(action)s"
-                        "%(destination)s%(unmodified)s.\n")
+            template = ("\n%(modified_count)s %(identifier)s %(action)s"
+                        "%(destination)s%(unmodified)s%(post_processed)s.\n")
             summary = template % {
-                'actual_count': actual_count,
-                'identifier': 'static file' + (actual_count > 1 and 's' or ''),
+                'modified_count': modified_count,
+                'identifier': 'static file' + (modified_count != 1 and 's' or ''),
                 'action': self.symlink and 'symlinked' or 'copied',
                 'destination': (destination_path and " to '%s'"
                                 % destination_path or ''),
-                'unmodified': (self.unmodified_files and ', %s unmodified'
+                'unmodified': (collected['unmodified'] and ', %s unmodified'
                                % unmodified_count or ''),
+                'post_processed': (collected['post_processed'] and
+                                   ', %s post-processed'
+                                   % post_processed_count or ''),
             }
             self.stdout.write(smart_str(summary))