+++ staticfiles/storage.py
@@ -3,23 +3,43 @@ import os
 import posixpath
 import re
 import warnings
+
 from datetime import datetime
+from urllib import unquote
+from urlparse import urlsplit, urlunsplit, urldefrag
 
 from django import VERSION
 from django.conf import settings
 from django.core.cache import (get_cache, InvalidCacheBackendError,
                                cache as default_cache)
-from django.core.files.base import ContentFile
+from django.core.files.base import File, ContentFile
 from django.core.files.storage import FileSystemStorage, get_storage_class
 from django.core.exceptions import ImproperlyConfigured
-from django.utils.encoding import force_unicode
+from django.utils.encoding import force_unicode, smart_str
 from django.utils.datastructures import SortedDict
 from django.utils.functional import LazyObject
 from django.utils.importlib import import_module
 from django.utils.hashcompat import md5_constructor
 
-
 from staticfiles.utils import matches_patterns
+
+
+def setattr_ifmissing(clss, name, func):
+    for cls in clss:
+        if not hasattr(cls, name):
+            setattr(cls, name, func)
+
+
+def __enter__(self):
+    return self
+
+
+def __exit__(self, exc_type, exc_value, tb):
+    self.close()
+
+# forcing the storage backend file to be a context manager
+setattr_ifmissing([File, ContentFile], '__enter__', __enter__)
+setattr_ifmissing([File, ContentFile], '__exit__', __exit__)
 
 
 class TimeAwareFileSystemStorage(FileSystemStorage):
@@ -95,23 +115,33 @@ class CachedFilesMixin(object):
                 self._patterns.setdefault(extension, []).append(compiled)
 
     def hashed_name(self, name, content=None):
+        parsed_name = urlsplit(unquote(name))
+        clean_name = parsed_name.path
         if content is None:
-            if not self.exists(name):
-                raise ValueError(
-                    "The file '%s' could not be found with %r." % (name, self))
+            if not self.exists(clean_name):
+                raise ValueError("The file '%s' could not be found with %r." %
+                                 (clean_name, self))
             try:
-                content = self.open(name)
+                content = self.open(clean_name)
             except IOError:
-                # Handle directory paths
+                # Handle directory paths and fragments
                 return name
-        path, filename = os.path.split(name)
+        path, filename = os.path.split(clean_name)
         root, ext = os.path.splitext(filename)
         # Get the MD5 hash of the file
         md5 = md5_constructor()
         for chunk in content.chunks():
             md5.update(chunk)
         md5sum = md5.hexdigest()[:12]
-        return os.path.join(path, u"%s.%s%s" % (root, md5sum, ext))
+        hashed_name = os.path.join(path, u"%s.%s%s" %
+                                   (root, md5sum, ext))
+        unparsed_name = list(parsed_name)
+        unparsed_name[2] = hashed_name
+        # Special casing for a @font-face hack, like url(myfont.eot?#iefix")
+        # http://www.fontspring.com/blog/the-new-bulletproof-font-face-syntax
+        if '?#' in name and not unparsed_name[3]:
+            unparsed_name[2] += '?'
+        return urlunsplit(unparsed_name)
 
     def cache_key(self, name):
         return u'staticfiles:cache:%s' % name
@@ -121,12 +151,31 @@ class CachedFilesMixin(object):
         Returns the real URL in DEBUG mode.
         """
         if settings.DEBUG and not force:
-            return super(CachedFilesMixin, self).url(name)
-        cache_key = self.cache_key(name)
-        hashed_name = self.cache.get(cache_key)
-        if hashed_name is None:
-            hashed_name = self.hashed_name(name)
-        return super(CachedFilesMixin, self).url(hashed_name)
+            hashed_name, fragment = name, ''
+        else:
+            clean_name, fragment = urldefrag(name)
+            cache_key = self.cache_key(name)
+            hashed_name = self.cache.get(cache_key)
+            if hashed_name is None:
+                hashed_name = self.hashed_name(clean_name).replace('\\', '/')
+                # set the cache if there was a miss
+                # (e.g. if cache server goes down)
+                self.cache.set(cache_key, hashed_name)
+
+        final_url = super(CachedFilesMixin, self).url(hashed_name)
+
+        # Special casing for a @font-face hack, like url(myfont.eot?#iefix")
+        # http://www.fontspring.com/blog/the-new-bulletproof-font-face-syntax
+        query_fragment = '?#' in name  # [sic!]
+        if fragment or query_fragment:
+            urlparts = list(urlsplit(final_url))
+            if fragment and not urlparts[4]:
+                urlparts[4] = fragment
+            if query_fragment and not urlparts[3]:
+                urlparts[2] += '?'
+            final_url = urlunsplit(urlparts)
+
+        return unquote(final_url)
 
     def url_converter(self, name, fail_silently=False):
         """
@@ -139,10 +188,11 @@ class CachedFilesMixin(object):
             of the storage.
             """
             matched, url = matchobj.groups()
-            # Completely ignore http(s) prefixed URLs
-            if url.startswith(('http', 'https')):
+            # Completely ignore http(s) prefixed URLs,
+            # fragments and data-uri URLs
+            if url.startswith(('#', 'http:', 'https:', 'data:')):
                 return matched
-            name_parts = name.split('/')
+            name_parts = name.split(os.sep)
             # Using posix normpath here to remove duplicates
             url = posixpath.normpath(url)
             url_parts = url.split('/')
@@ -161,65 +211,86 @@ class CachedFilesMixin(object):
                     start, end = 1, sub_level - 1
             joined_result = '/'.join(name_parts[:-start] + url_parts[end:])
             try:
-                hashed_url = self.url(joined_result, force=True)
+                hashed_url = self.url(unquote(joined_result), force=True)
             except ValueError:
                 if not fail_silently:
                     raise
                 hashed_url = url
             # Return the hashed and normalized version to the file
-            return 'url("%s")' % hashed_url
+            return 'url("%s")' % unquote(hashed_url)
         return converter
 
     def post_process(self, paths, dry_run=False, **options):
         """
         Post process the given list of files (called from collectstatic).
+
+        Processing is actually two separate operations:
+
+        1. renaming files to include a hash of their content for cache-busting,
+           and copying those files to the target storage.
+        2. adjusting files which contain references to other files so they
+           refer to the cache-busting filenames.
+
+        If either of these are performed on a file, then that file is considered
+        post-processed.
         """
         fail_silently = options['ignore_errors']
-        processed_files = []
         # don't even dare to process the files if we're in dry run mode
         if dry_run:
-            return processed_files
+            return
 
         # delete cache of all handled paths
         self.cache.delete_many([self.cache_key(path) for path in paths])
 
-        # only try processing the files we have patterns for
+        # build a list of adjustable files
         matches = lambda path: matches_patterns(path, self._patterns.keys())
-        processing_paths = [path for path in paths if matches(path)]
+        adjustable_paths = [path for path in paths if matches(path)]
 
         # then sort the files by the directory level
         path_level = lambda name: len(name.split(os.sep))
-        for name in sorted(paths, key=path_level, reverse=True):
+        for name in sorted(paths.keys(), key=path_level, reverse=True):
 
-            # first get a hashed name for the given file
-            hashed_name = self.hashed_name(name)
+            # use the original, local file, not the copied-but-unprocessed
+            # file, which might be somewhere far away, like S3
+            storage, path = paths[name]
+            with storage.open(path) as original_file:
 
-            original_file = self.open(name)
-            try:
-                # then get the original's file content
-                content = original_file.read()
+                # generate the hash with the original content, even for
+                # adjustable files.
+                hashed_name = self.hashed_name(name, original_file)
 
-                # to apply each replacement pattern on the content
-                if name in processing_paths:
+                # then get the original's file content..
+                if hasattr(original_file, 'seek'):
+                    original_file.seek(0)
+
+                hashed_file_exists = self.exists(hashed_name)
+                processed = False
+
+                # ..to apply each replacement pattern to the content
+                if name in adjustable_paths:
+                    content = original_file.read()
                     converter = self.url_converter(name, fail_silently)
                     for patterns in self._patterns.values():
                         for pattern in patterns:
                             content = pattern.sub(converter, content)
-
-                # then save the processed result
-                if self.exists(hashed_name):
-                    self.delete(hashed_name)
-
-                saved_name = self._save(hashed_name, ContentFile(content))
-                hashed_name = force_unicode(saved_name.replace('\\', '/'))
-                processed_files.append(hashed_name)
+                    if hashed_file_exists:
+                        self.delete(hashed_name)
+                    # then save the processed result
+                    content_file = ContentFile(smart_str(content))
+                    saved_name = self._save(hashed_name, content_file)
+                    hashed_name = force_unicode(saved_name.replace('\\', '/'))
+                    processed = True
+                else:
+                    # or handle the case in which neither processing nor
+                    # a change to the original file happened
+                    if not hashed_file_exists:
+                        processed = True
+                        saved_name = self._save(hashed_name, original_file)
+                        hashed_name = force_unicode(saved_name.replace('\\', '/'))
 
                 # and then set the cache accordingly
                 self.cache.set(self.cache_key(name), hashed_name)
-            finally:
-                original_file.close()
-
-        return processed_files
+                yield name, hashed_name, processed
 
 
 class CachedStaticFilesStorage(CachedFilesMixin, StaticFilesStorage):