+++ staticfiles/tests/tests.py
@@ -7,21 +7,15 @@ import posixpath
 import shutil
 import sys
 import tempfile
+import unittest2
 from StringIO import StringIO
 
-try:
-    from cStringIO import StringIO
-except ImportError:
-    from StringIO import StringIO
-
-from django import VERSION
-from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 from django.core.files.storage import default_storage
 from django.core.management import call_command
 from django.template import loader, Context
 from django.test import TestCase
-from django.utils.encoding import smart_unicode, smart_str, force_unicode
+from django.utils.encoding import smart_unicode
 
 try:
     from django.utils.functional import empty
@@ -29,21 +23,9 @@ except ImportError:
     empty = None
 
 from staticfiles import finders, storage
-
-
-# This is the most annoying thing ever, but cStringIO and StringIO don't
-# return the same for Unicode strings, see http://bugs.python.org/issue2387
-# and http://tuxpool.blogspot.com/2010/01/python-cstringio-vs-stringio.html
-# For some reason this only shows up in Django < 1.3 WTF!
-is_fucking_old = VERSION[:2] < (1, 3)
-def u(string):
-    if is_fucking_old:
-        return StringIO(unicode(string)).read()
-    return string
-
-
-
-TEST_ROOT = os.path.normcase(os.path.dirname(__file__))
+from staticfiles.conf import settings
+from staticfiles.management.commands.collectstatic import Command as \
+    CollectstaticCommand
 
 
 def rmtree_errorhandler(func, path, exc_info):
@@ -80,19 +62,30 @@ class BaseStaticFilesTestCase(object):
         self.old_debug = settings.DEBUG
         settings.DEBUG = True
 
+        testfiles_path = os.path.join(settings.TEST_ROOT, 'apps', 'test', 'static', 'test')
         # To make sure SVN doesn't hangs itself with the non-ASCII characters
         # during checkout, we actually create one file dynamically.
-        self._nonascii_filepath = os.path.join(
-            TEST_ROOT, 'apps', 'test', 'static', 'test', u'fi\u015fier.txt')
+        self._nonascii_filepath = os.path.join(testfiles_path, u'fi\u015fier.txt')
         with codecs.open(self._nonascii_filepath, 'w', 'utf-8') as f:
             f.write(u"fi\u015fier in the app dir")
+        # And also create the stupid hidden file to dwarf the setup.py's
+        # package data handling.
+        self._hidden_filepath = os.path.join(testfiles_path, '.hidden')
+        with codecs.open(self._hidden_filepath, 'w', 'utf-8') as f:
+            f.write("should be ignored")
+        self._backup_filepath = os.path.join(
+            settings.TEST_ROOT, 'project', 'documents', 'test', 'backup~')
+        with codecs.open(self._backup_filepath, 'w', 'utf-8') as f:
+            f.write("should be ignored")
 
     def tearDown(self):
         os.unlink(self._nonascii_filepath)
+        os.unlink(self._hidden_filepath)
+        os.unlink(self._backup_filepath)
         settings.DEBUG = self.old_debug
 
     def assertFileContains(self, filepath, text):
-        self.assertTrue(text in self._get_file(smart_unicode(filepath)),
+        self.assertIn(text, self._get_file(smart_unicode(filepath)),
                         u"'%s' not in '%s'" % (text, filepath))
 
     def assertFileNotFound(self, filepath):
@@ -103,18 +96,22 @@ class BaseStaticFilesTestCase(object):
             template = loader.get_template_from_string(template)
         return template.render(Context(kwargs)).strip()
 
-    def assertTemplateRenders(self, template, result, **kwargs):
+    def static_template_snippet(self, path):
+        return "{%% load staticfiles %%}{%% static '%s' %%}" % path
+
+    def assertStaticRenders(self, path, result, **kwargs):
+        template = self.static_template_snippet(path)
         self.assertEqual(self.render_template(template, **kwargs), result)
 
-    def assertTemplateRaises(self, exc, template, result, **kwargs):
-        self.assertRaises(exc, self.assertTemplateRenders, template, result, **kwargs)
+    def assertStaticRaises(self, exc, path, result, **kwargs):
+        self.assertRaises(exc, self.assertStaticRenders, path, result, **kwargs)
 
 
-class StaticFilesTestCase(BaseStaticFilesTestCase, TestCase):
+class StaticFilesTestCase(BaseStaticFilesTestCase, TestCase, unittest2.TestCase):
     pass
 
 
-class BaseCollectionTestCase(BaseStaticFilesTestCase):
+class BaseCollectionTestCase(BaseStaticFilesTestCase, unittest2.TestCase):
     """
     Tests shared by all file finding features (collectstatic,
     findstatic, and static serve view).
@@ -126,7 +123,7 @@ class BaseCollectionTestCase(BaseStaticFilesTestCase):
     def setUp(self):
         super(BaseCollectionTestCase, self).setUp()
         self.old_root = settings.STATIC_ROOT
-        settings.STATIC_ROOT = tempfile.mkdtemp()
+        settings.STATIC_ROOT = tempfile.mkdtemp(prefix='staticfiles_')
         self.run_collectstatic()
 
     def tearDown(self):
@@ -136,14 +133,18 @@ class BaseCollectionTestCase(BaseStaticFilesTestCase):
         super(BaseCollectionTestCase, self).tearDown()
 
     def run_collectstatic(self, **kwargs):
+        ignore_patterns = [
+            '*.ignoreme', os.path.join('test', '*.ignoreme2'), os.path.join(
+                settings.TEST_ROOT, 'apps', 'test', 'static', 'test', '*.ignoreme3')]
         call_command('collectstatic', interactive=False, verbosity='0',
-                     ignore_patterns=['*.ignoreme'], **kwargs)
+                     ignore_patterns=ignore_patterns, **kwargs)
 
     def _get_file(self, filepath):
         assert filepath, 'filepath is empty.'
         filepath = os.path.join(settings.STATIC_ROOT, filepath)
         with codecs.open(filepath, "r", "utf-8") as f:
             return f.read()
+
 
 class CollectionTestCase(BaseCollectionTestCase, StaticFilesTestCase):
     pass
@@ -243,6 +244,8 @@ class TestCollection(CollectionTestCase, TestDefaults):
         Test that -i patterns are ignored.
         """
         self.assertFileNotFound('test/test.ignoreme')
+        self.assertFileNotFound('test/test_relative.ignoreme2')
+        self.assertFileNotFound('test/test_absolute.ignoreme3')
 
     def test_common_ignore_patterns(self):
         """
@@ -251,6 +254,16 @@ class TestCollection(CollectionTestCase, TestDefaults):
         self.assertFileNotFound('test/.hidden')
         self.assertFileNotFound('test/backup~')
         self.assertFileNotFound('test/CVS')
+
+    def test_staticfiles_ignore_patterns(self):
+        """
+        Test that patterns in STATICFILES_IGNORE_PATTERNS are ignored.
+        """
+        self.assertFileNotFound('test/test.ignoreme4')
+        self.assertFileNotFound('test/test_relative.ignoreme5')
+        self.assertFileNotFound('test/test_absolute.ignoreme6')
+        self.assertFileNotFound('prefix/test.ignoreme7')
+        self.assertFileNotFound('ignored/test_directory.txt')
 
 
 class TestCollectionClear(CollectionTestCase):
@@ -310,7 +323,7 @@ class TestCollectionNonLocalStorage(CollectionTestCase, TestNoFilesCreated):
     """
     def setUp(self):
         self.old_staticfiles_storage = settings.STATICFILES_STORAGE
-        settings.STATICFILES_STORAGE = 'tests.storage.DummyStorage'
+        settings.STATICFILES_STORAGE = 'staticfiles.tests.storage.DummyStorage'
         super(TestCollectionNonLocalStorage, self).setUp()
 
     def tearDown(self):
@@ -344,86 +357,136 @@ class TestCollectionCachedStorage(BaseCollectionTestCase, BaseStaticFilesTestCas
         """
         Test the CachedStaticFilesStorage backend.
         """
-        self.assertTemplateRaises(ValueError, """
-            {% load staticfiles %}{% static "does/not/exist.png" %}
-            """, "/static/does/not/exist.png")
-        self.assertTemplateRenders("""
-            {% load staticfiles %}{% static "test/file.txt" %}
-            """, "/static/test/file.dad0999e4f8f.txt")
-        self.assertTemplateRenders("""
-            {% load staticfiles %}{% static "cached/styles.css" %}
-            """, "/static/cached/styles.93b1147e8552.css")
+        self.assertStaticRaises(ValueError,
+                                "does/not/exist.png",
+                                "/static/does/not/exist.png")
+        self.assertStaticRenders("test/file.txt",
+                                 "/static/test/file.ea5bccaf16d5.txt")
+        self.assertStaticRenders("cached/styles.css",
+                                 "/static/cached/styles.93b1147e8552.css")
 
     def test_template_tag_simple_content(self):
         relpath = self.cached_file_path("cached/styles.css")
         self.assertEqual(relpath, "cached/styles.93b1147e8552.css")
-        relfile = storage.staticfiles_storage.open(relpath)
-        try:
+        with storage.staticfiles_storage.open(relpath) as relfile:
             content = relfile.read()
-            self.assertFalse(u("cached/other.css") in content, content)
-            self.assertTrue(u("/static/cached/other.d41d8cd98f00.css") in content)
-        finally:
-            relfile.close()
+            self.assertNotIn("cached/other.css", content)
+            self.assertIn("/static/cached/other.d41d8cd98f00.css", content)
+
+    def test_path_with_querystring(self):
+        relpath = self.cached_file_path("cached/styles.css?spam=eggs")
+        self.assertEqual(relpath,
+                         "cached/styles.93b1147e8552.css?spam=eggs")
+        with storage.staticfiles_storage.open(
+                "cached/styles.93b1147e8552.css") as relfile:
+            content = relfile.read()
+            self.assertNotIn("cached/other.css", content)
+            self.assertIn("/static/cached/other.d41d8cd98f00.css", content)
+
+    def test_path_with_fragment(self):
+        relpath = self.cached_file_path("cached/styles.css#eggs")
+        self.assertEqual(relpath, "cached/styles.93b1147e8552.css#eggs")
+        with storage.staticfiles_storage.open(
+                "cached/styles.93b1147e8552.css") as relfile:
+            content = relfile.read()
+            self.assertNotIn("cached/other.css", content)
+            self.assertIn("/static/cached/other.d41d8cd98f00.css", content)
+
+    def test_path_with_querystring_and_fragment(self):
+        relpath = self.cached_file_path("cached/css/fragments.css")
+        self.assertEqual(relpath, "cached/css/fragments.75433540b096.css")
+        with storage.staticfiles_storage.open(relpath) as relfile:
+            content = relfile.read()
+            self.assertIn('/static/cached/css/fonts/font.a4b0478549d0.eot?#iefix', content)
+            self.assertIn('/static/cached/css/fonts/font.b8d603e42714.svg#webfontIyfZbseF', content)
+            self.assertIn('data:font/woff;charset=utf-8;base64,d09GRgABAAAAADJoAA0AAAAAR2QAAQAAAAAAAAAAAAA', content)
+            self.assertIn('#default#VML', content)
 
     def test_template_tag_absolute(self):
         relpath = self.cached_file_path("cached/absolute.css")
         self.assertEqual(relpath, "cached/absolute.cc80cb5e2eb1.css")
-        relfile = storage.staticfiles_storage.open(relpath)
-        try:
+        with storage.staticfiles_storage.open(relpath) as relfile:
             content = relfile.read()
-            self.assertFalse(u("/static/cached/styles.css") in content)
-            self.assertTrue(u("/static/cached/styles.93b1147e8552.css") in content)
-        finally:
-            relfile.close()
+            self.assertNotIn("/static/cached/styles.css", content)
+            self.assertIn("/static/cached/styles.93b1147e8552.css", content)
 
     def test_template_tag_denorm(self):
         relpath = self.cached_file_path("cached/denorm.css")
         self.assertEqual(relpath, "cached/denorm.363de96e9b4b.css")
-        relfile = storage.staticfiles_storage.open(relpath)
-        try:
+        with storage.staticfiles_storage.open(relpath) as relfile:
             content = relfile.read()
-            self.assertFalse(u("..//cached///styles.css") in content)
-            self.assertTrue(u("/static/cached/styles.93b1147e8552.css") in content)
-        finally:
-            relfile.close()
+            self.assertNotIn("..//cached///styles.css", content)
+            self.assertIn("/static/cached/styles.93b1147e8552.css", content)
 
     def test_template_tag_relative(self):
         relpath = self.cached_file_path("cached/relative.css")
         self.assertEqual(relpath, "cached/relative.8dffb45d91f5.css")
-        relfile = storage.staticfiles_storage.open(relpath)
-        try:
+        with storage.staticfiles_storage.open(relpath) as relfile:
             content = relfile.read()
-            self.assertTrue(u("/static/cached/styles.93b1147e8552.css") in content)
-            self.assertFalse(u("../cached/styles.css") in content)
-            self.assertFalse(u('@import "styles.css"') in content)
-            self.assertFalse(u('url(img/relative.png)') in content)
-            self.assertTrue(u('url("/static/cached/img/relative.acae32e4532b.png")') in content)
-            self.assertTrue(u("/static/cached/styles.93b1147e8552.css") in content)
-        finally:
-            relfile.close()
+            self.assertIn("/static/cached/styles.93b1147e8552.css", content)
+            self.assertNotIn("../cached/styles.css", content)
+            self.assertNotIn('@import "styles.css"', content)
+            self.assertNotIn('url(img/relative.png)', content)
+            self.assertIn('url("/static/cached/img/relative.acae32e4532b.png")', content)
+            self.assertIn("/static/cached/styles.93b1147e8552.css", content)
 
     def test_template_tag_deep_relative(self):
         relpath = self.cached_file_path("cached/css/window.css")
         self.assertEqual(relpath, "cached/css/window.9db38d5169f3.css")
-        relfile = storage.staticfiles_storage.open(relpath)
-        try:
+        with storage.staticfiles_storage.open(relpath) as relfile:
             content = relfile.read()
-            self.assertFalse(u('url(img/window.png)') in content)
-            self.assertTrue(u('url("/static/cached/css/img/window.acae32e4532b.png")') in content)
-        finally:
-            relfile.close()
+            self.assertNotIn('url(img/window.png)', content)
+            self.assertIn('url("/static/cached/css/img/window.acae32e4532b.png")', content)
 
     def test_template_tag_url(self):
         relpath = self.cached_file_path("cached/url.css")
         self.assertEqual(relpath, "cached/url.615e21601e4b.css")
-        relfile = storage.staticfiles_storage.open(relpath)
-        try:
-            self.assertTrue("https://" in relfile.read())
-        finally:
-            relfile.close()
+        with storage.staticfiles_storage.open(relpath) as relfile:
+            self.assertIn("https://", relfile.read())
 
+    def test_cache_invalidation(self):
+        name = "cached/styles.css"
+        hashed_name = "cached/styles.93b1147e8552.css"
+        # check if the cache is filled correctly as expected
+        cache_key = storage.staticfiles_storage.cache_key(name)
+        cached_name = storage.staticfiles_storage.cache.get(cache_key)
+        self.assertEqual(self.cached_file_path(name), cached_name)
+        # clearing the cache to make sure we re-set it correctly in the url method
+        storage.staticfiles_storage.cache.clear()
+        cached_name = storage.staticfiles_storage.cache.get(cache_key)
+        self.assertEqual(cached_name, None)
+        self.assertEqual(self.cached_file_path(name), hashed_name)
+        cached_name = storage.staticfiles_storage.cache.get(cache_key)
+        self.assertEqual(cached_name, hashed_name)
+
+    def test_post_processing(self):
+        """Test that post_processing behaves correctly.
+
+        Files that are alterable should always be post-processed; files that
+        aren't should be skipped.
+
+        collectstatic has already been called once in setUp() for this testcase,
+        therefore we check by verifying behavior on a second run.
+        """
+        collectstatic_args = {
+            'interactive': False,
+            'verbosity': '0',
+            'link': False,
+            'clear': False,
+            'dry_run': False,
+            'post_process': True,
+            'use_default_ignore_patterns': True,
+            'ignore_patterns': ['*.ignoreme']
+        }
+
+        collectstatic_cmd = CollectstaticCommand()
+        collectstatic_cmd.set_options(**collectstatic_args)
+        stats = collectstatic_cmd.collect()
+        self.assertTrue(u'cached/css/window.css' in stats['post_processed'])
+        self.assertTrue(u'cached/css/img/window.png' in stats['unmodified'])
 
 if sys.platform != 'win32':
+
     class TestCollectionLinks(CollectionTestCase, TestDefaults):
         """
         Test ``--link`` option for ``collectstatic`` management command.
@@ -446,7 +509,7 @@ class TestServeStatic(StaticFilesTestCase):
     """
     Test static asset serving view.
     """
-    urls = 'tests.urls.default'
+    urls = 'staticfiles.tests.urls.default'
 
     def _response(self, filepath):
         return self.client.get(
@@ -482,11 +545,12 @@ class TestServeStaticWithDefaultURL(TestServeStatic, TestDefaults):
     """
     pass
 
+
 class TestServeStaticWithURLHelper(TestServeStatic, TestDefaults):
     """
     Test static asset serving view with staticfiles_urlpatterns helper.
     """
-    urls = 'tests.urls.helper'
+    urls = 'staticfiles.tests.urls.helper'
 
 
 class TestServeAdminMedia(TestServeStatic):
@@ -529,7 +593,7 @@ class TestFileSystemFinder(StaticFilesTestCase, FinderTestCase):
     def setUp(self):
         super(TestFileSystemFinder, self).setUp()
         self.finder = finders.FileSystemFinder()
-        test_file_path = os.path.join(TEST_ROOT, 'project', 'documents', 'test', 'file.txt')
+        test_file_path = os.path.join(settings.TEST_ROOT, 'project', 'documents', 'test', 'file.txt')
         self.find_first = (os.path.join('test', 'file.txt'), test_file_path)
         self.find_all = (os.path.join('test', 'file.txt'), [test_file_path])
 
@@ -541,7 +605,7 @@ class TestAppDirectoriesFinder(StaticFilesTestCase, FinderTestCase):
     def setUp(self):
         super(TestAppDirectoriesFinder, self).setUp()
         self.finder = finders.AppDirectoriesFinder()
-        test_file_path = os.path.join(TEST_ROOT, 'apps', 'test', 'static', 'test', 'file1.txt')
+        test_file_path = os.path.join(settings.TEST_ROOT, 'apps', 'test', 'static', 'test', 'file1.txt')
         self.find_first = (os.path.join('test', 'file1.txt'), test_file_path)
         self.find_all = (os.path.join('test', 'file1.txt'), [test_file_path])
 
@@ -592,9 +656,6 @@ class TestMiscFinder(TestCase):
 class TestTemplateTag(StaticFilesTestCase):
 
     def test_template_tag(self):
-        self.assertTemplateRenders("""
-            {% load staticfiles %}{% static "does/not/exist.png" %}
-            """, "/static/does/not/exist.png")
-        self.assertTemplateRenders("""
-            {% load staticfiles %}{% static "testfile.txt" %}
-            """, "/static/testfile.txt")
+        self.assertStaticRenders("does/not/exist.png",
+                                   "/static/does/not/exist.png")
+        self.assertStaticRenders("testfile.txt", "/static/testfile.txt")