+++ docs/changelog.rst
@@ -1,6 +1,26 @@
 Changelog
 =========
 
+v1.2.1 (2012-02-16)
+-------------------
+
+* Backported a change from Django trunk that prevents openening too
+  many files at once when running the ``collectstatic`` management
+  command.
+
+v1.2 (2012-02-12)
+-----------------
+
+* Added ``STATICFILES_IGNORE_PATTERNS`` setting to globally ignore
+  files when running the ``collectstatic`` management command.
+
+* Refactored ``CachedFilesMixin`` and management command to only
+  post process the collected files if really needed.
+
+* Added support for URL fragment to the ``CachedStaticFilesStorage``.
+
+* Stopped using versiontools_ again as it caused installation time issues.
+
 v1.1.2 (2011-08-25)
 -------------------
 
@@ -11,10 +31,10 @@ v1.1.1 (2011-08-22)
 
 * Fixed resolution of relative paths in ``CachedStaticFilesStorage``.
 
-* Started to use `django-appconf`_ and `versiontools`_.
+* Started to use `django-appconf`_ and versiontools_.
 
 .. _`django-appconf`: http://django-appconf.rtfd.org/
-.. _`versiontools`: http://pypi.python.org/pypi/versiontools
+.. _versiontools: http://pypi.python.org/pypi/versiontools
 
 v1.1 (2011-08-18)
 -----------------